move, rename, and namespace transports
