+++ lib/lightstep/tracer/transport/callback.rb
@@ -1,0 +1,17 @@
+module LightStep
+  module Transport
+    class Callback
+      def initialize(callback:)
+        @callback = callback
+      end
+
+      def flush_report(_auth, report)
+        @callback.call(report)
+        nil
+      end
+
+      def close(immediate)
+      end
+    end
+  end
+end
